+++ packages/client/src/components/Players/Opponent.tsx
@@ -3,7 +3,7 @@ import { CardWrapper } from 'components/Cards/CardWrapper';
 import { Deck } from 'components/Cards/Deck';
 import { PlayerState, cardKey } from 'shared';
 import { Box, BoxProps, Grid, Text } from 'theme-ui';
-import { PLAYER_DECK_ID } from 'utils/dom';
+import { playerDeckId } from 'utils/dom';
 
 type Props = {
   player: PlayerState;
@@ -15,15 +15,11 @@ export const Opponent = ({ player, activePlayerCard, ...rest }: Props) => {
     return <Text>Waiting for a player to join</Text>;
   }
   const { captured, hand, username, scopa } = player;
+  
   return (
     <Box {...rest}>
       <Grid sx={{ m: 3, marginTop: '-7vw' }} columns="1.5fr 1fr 1fr 1fr">
-        <Deck
-          cardNumber={captured.length}
-          scopa={scopa}
-          id={`${PLAYER_DECK_ID}__${username}`}
-          scopaPosition="bottom"
-        />
+        <Deck cardNumber={captured.length} scopa={scopa} id={playerDeckId(username)} scopaPosition="bottom" />
         {hand?.map((c) => {
           const key = cardKey(c);
           return (
